update ood-portal-generator and fix mod_ood_proxy
