+++ mod_ood_proxy/lib/analytics.lua
@@ -1,26 +1,3 @@
---[[
-  table_to_json
-
-  A simple helper function to turn a lua table into a json object.
---]]
-function table_to_json(lua_table)
-  local result = {}
-
-  for key, value in pairs(lua_table) do
-      table.insert(result, string.format("\"%s\": \"%s\"", key, value))
-  end
-
-  return "{" .. table.concat(result, ",") .. "}"
-end
-
---[[
-  ga_body
-
-  Google Analytics body - a simple helper function to generate json data for the GA API.
---]]
-function ga_body(client_id, event_data)
-  return string.format("{ \"client_id\": \"%s\", \"events\": [%s] }", client_id, event_data))
-end
 
 --[[
   analytics_handler
@@ -29,8 +6,10 @@ end
   phoning home with analytics (required by OOD proposal).
 --]]
 function analytics_handler(r)
+  local json = require 'ood.json'
+
   -- read in OOD specific settings defined in the Apache config
-  local url = 'www.google-analytics.com/mp/collect'
+  local url = 'https://www.google-analytics.com/mp/collect'
   local tracking_id  = r.subprocess_env['OOD_GA_TRACKING_ID'] or r.subprocess_env['OOD_ANALYTICS_TRACKING_ID']
   local api_key      = r.subprocess_env['OOD_GA_API_SECRET']
 
@@ -61,13 +40,15 @@ function analytics_handler(r)
     data['path']        = r.uri
     data['method']      = r.method
 
-    post_body = ga_body(client_id, data)
+    local now = r:clock()
+    event_data = json.table_to_json(data)
+    post_body = json.ga_body(client_id, event_data)
     full_url = string.format("%s?measurement_id=%s&api_secret=%s", url, tracking_id, api_key)
 
-    local handle = io.popen("wget --post-data='" .. post_body .. "' " .. full_url .. " -O /dev/null -T 5 -nv 2>&1")
+    local handle = io.popen("wget --header='Content-type: application/json' --post-data='" .. post_body .. "' '" .. full_url .. "' -O /dev/null -T 5 -nv 2>&1")
     output = handle:read('*all'):match('^%s*(.-)%s*$')
     handle:close()
-    r:debug("Analytics input: '" .. query .. "'")
+    r:debug("Analytics input: '" .. event_data .. "'")
     r:debug("Analytics output: '" .. r:escape_logitem(output) .. "' [" .. (r:clock() - now)/1000.0 .. " ms]")
   end
