+++ ood-portal-generator/share/ood_portal_example.yml
@@ -167,7 +167,7 @@
 # information on how to setup the GA property)
 # Example:
 #     analytics:
-#       url: 'http://www.google-analytics.com/collect'
+#       secret: 'my-google-analytics-secret'
 #       id: 'UA-79331310-4'
 # Default: null (do not track)
 #analytics: null
