+++ ood-portal-generator/spec/fixtures/ood-portal.conf.all
@@ -198,8 +198,8 @@ Listen 8080
     SetEnv OOD_PUN_MAX_RETRIES "3000"
     LuaHookFixups pun_proxy.lua pun_proxy_handler
 
-    SetEnv OOD_ANALYTICS_TRACKING_URL "http://www.google-analytics.com/collect"
-    SetEnv OOD_ANALYTICS_TRACKING_ID  "analytics-id"
+    SetEnv OOD_GA_API_SECRET "analytics-api-secret"
+    SetEnv OOD_GA_TRACKING_ID  "analytics-id"
     LuaHookLog analytics.lua analytics_handler
   </Location>
