+++ ood-portal-generator/spec/fixtures/ood_portal.yaml.all
@@ -162,11 +162,11 @@ root_uri: '/other/dashboard'
 # information on how to setup the GA property)
 # Example:
 #     analytics:
-#       url: 'http://www.google-analytics.com/collect'
+#       secret: 'analytics-api-secret'
 #       id: 'UA-79331310-4'
 # Default: null (do not track)
 analytics:
-  url: 'http://www.google-analytics.com/collect'
+  secret: 'analytics-api-secret'
   id: 'analytics-id'
 
 #
